Update README.md
The purpose of this project was to analyze how traffic (Bus, Car, Pedestrians, Cycles) patterns change throughout the day and throughout the semester and find anomalies and then to come up with a suggestion to improve the traffic condition. I performed Machine Learning algorithms to predict future traffic conditions.
The Data set is collected the Department of Transportation at University of Maryland College Park.